Fixed unfortunate usize

+++ yave/graphics/device/PhysicalDevice.cpp
@@ -85,8 +85,8 @@ bool PhysicalDevice::is_discrete() const {
     return vk_properties().deviceType == VK_PHYSICAL_DEVICE_TYPE_DISCRETE_GPU;
 }
 
-usize PhysicalDevice::total_device_memory() const {
-    usize total = 0;
+u64 PhysicalDevice::total_device_memory() const {
+    u64 total = 0;
     for(u32 i = 0; i != _memory_properties.memoryHeapCount; ++i) {
         if(_memory_properties.memoryHeaps[i].flags & VK_MEMORY_HEAP_DEVICE_LOCAL_BIT) {
             total += _memory_properties.memoryHeaps[i].size;
